Dev: Moved new form classes in existing form/base directory.
